+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_passing_async_when_sync_expected/avoid_passing_async_when_sync_expected_rule.dart
@@ -36,10 +36,7 @@ class AvoidPassingAsyncWhenSyncExpectedRule extends CommonRule {
     return visitor.invalidArguments
         .map((invocation) => createIssue(
               rule: this,
-              location: nodeLocation(
-                node: invocation,
-                source: source,
-              ),
+              location: nodeLocation(node: invocation, source: source),
               message: _warningMessage,
             ))
         .toList(growable: false);